Add Contract Channel Method call and Move BranchContract package
