+++ yggdrash-common/src/main/java/io/yggdrash/common/store/BranchStateStore.java
@@ -17,6 +17,7 @@
 package io.yggdrash.common.store;
 
 import io.yggdrash.common.Sha3Hash;
+import io.yggdrash.common.contract.BranchContract;
 import io.yggdrash.common.contract.vo.dpoa.ValidatorSet;
 import java.util.List;
